feat: add support for subscription syncing via LibreTube sync server
